+++ pwm_servo.h
@@ -1,7 +1,7 @@
 #ifndef PWM3_H_
 #define PWM3_H_
 
-void SetDutyCycleServo(unsigned int DutyCycle, unsigned int Frequency);
+void SetDutyCycleServo(float DutyCycle, unsigned int Frequency);
 void InitPWM3(void);
 //void FTM3_IRQHandler(void);
